+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add push/pull files and folder: `driver.device.pushFile`, `driver.device.pullFile`, `driver.device.pullFolder`
 - Add set/get clipboard: `driver.device.setClipboard`, `driver.device.getClipboard`
 - Add open notification: `driver.device.openNotification`
+- Add start activity: `driver.device.startActivity`
 
 ## [0.0.6] - 2019-06-30
 - Add device lock related commands: `driver.device.*`